requirements.txt
appnope==0.1.3
asttokens==2.4.0
backcall==0.2.0
bagpy==0.5
bitarray==2.8.1
bitstring==4.1.2
boto==2.49.0
boto3==1.28.53
botocore==1.31.53
catkin-pkg==1.0.0
catkin-tools==0.9.4
catkin-tools-python==0.1.1
certifi==2023.7.22
contourpy==1.1.1
cvbridge3==1.1
cycler==0.11.0
decorator==5.1.1
distro==1.8.0
docutils==0.20.1
empy==3.3.4
executing==1.2.0
fonttools==4.42.1
gnupg==2.3.1
importlib-resources==6.1.0
iniconfig==2.0.0
ipython==8.15.0
jedi==0.19.0
Jinja2==3.0.3
jmespath==1.0.1
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
minio==7.1.17
numpy==1.26.0
opencv-python==4.8.0.76
osrf-pycommon==2.0.2
packaging==23.1
pandas==2.1.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
pkginfo==1.9.6
pluggy==1.3.0
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py3rosmsgs==1.18.2
pycryptodomex==3.19.0
Pygments==2.16.1
pyparsing==3.1.1
pyserial==3.5
pytest==7.4.2
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
rospkg==1.5.0
s3transfer==0.6.2
seaborn==0.12.2
setuptools-scm==8.0.3
six==1.16.0
stack-data==0.6.2
traitlets==5.10.0
tzdata==2023.3
urllib3==1.26.16
wcwidth==0.2.6
